exercise 2.14 added window.confirm and exercise 2.15

+++ part2/phonebook/src/Persons.js
@@ -2,14 +2,20 @@ import personService from "./services/person"
 
 const Persons = ({onDelete, persons}) => {
 
-    const handleDelete = (id) => { 
+    const handleDelete = (id,name) => { 
+
         const idToDelete = id
-        onDelete(idToDelete)
+        const confirmDelete = window.confirm(`Delete ${name} ?`)
+        if (confirmDelete)
+        {
+            onDelete(idToDelete)
+        }
+        
     }
 
     return (
       <div>
-        {persons.map((person) => {return <p key = {person.name}>{person.name} {person.number}<button onClick = {() => handleDelete(person.id)}>delete</button></p>})}
+        {persons.map((person) => {return <p key = {person.name}>{person.name} {person.number}<button onClick = {() => handleDelete(person.id, person.name)}>delete</button></p>})}
       </div>
     )
   }
